created the validator coordinator to manage validators and make room to integrate the new p2p stuff
