+++ settings.gradle
@@ -16,5 +16,6 @@ include 'util'
 include 'errorprone-checks'
 include 'pow'
 include 'storage'
-include 'validatorclient'
+include 'validator:client'
+include 'validator:coordinator'
 include 'eth-reference-tests'
