+++ storage/src/main/java/tech/pegasys/artemis/storage/ChainStorageClient.java
@@ -21,8 +21,8 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Optional;
-import java.util.PriorityQueue;
 import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.PriorityBlockingQueue;
 import net.consensys.cava.bytes.Bytes;
 import org.apache.logging.log4j.Level;
 import tech.pegasys.artemis.datastructures.blocks.BeaconBlock;
@@ -35,10 +35,11 @@ import tech.pegasys.artemis.util.alogger.ALogger;
 /** This class is the ChainStorage client-side logic */
 public class ChainStorageClient implements ChainStorage {
   static final ALogger LOG = new ALogger(ChainStorageClient.class.getName());
-
+  static final Integer UNPROCESSED_BLOCKS_LENGTH = 100;
   protected final HashMap<Integer, Attestation> latestAttestations = new HashMap<>();
-  protected final PriorityQueue<BeaconBlock> unprocessedBlocks =
-      new PriorityQueue<>(Comparator.comparing(BeaconBlock::getSlot));
+  protected final PriorityBlockingQueue<BeaconBlock> unprocessedBlocks =
+      new PriorityBlockingQueue<BeaconBlock>(
+          UNPROCESSED_BLOCKS_LENGTH, Comparator.comparing(BeaconBlock::getSlot));
   protected final LinkedBlockingQueue<Attestation> unprocessedAttestations =
       new LinkedBlockingQueue<>();
   protected final HashMap<Bytes, BeaconBlock> processedBlockLookup = new HashMap<>();
@@ -82,7 +83,8 @@ public class ChainStorageClient implements ChainStorage {
    * @param block
    */
   public void addUnprocessedBlock(BeaconBlock block) {
-    ChainStorage.<BeaconBlock, PriorityQueue<BeaconBlock>>add(block, this.unprocessedBlocks);
+    ChainStorage.<BeaconBlock, PriorityBlockingQueue<BeaconBlock>>add(
+        block, this.unprocessedBlocks);
   }
 
   /**
@@ -139,11 +141,13 @@ public class ChainStorageClient implements ChainStorage {
     Optional<BeaconBlock> currentBlock;
     while (unproccesedBlocksLeft) {
       currentBlock =
-          ChainStorage.<BeaconBlock, PriorityQueue<BeaconBlock>>peek(this.unprocessedBlocks);
+          ChainStorage.<BeaconBlock, PriorityBlockingQueue<BeaconBlock>>peek(
+              this.unprocessedBlocks);
       if (currentBlock.isPresent()
           && UnsignedLong.valueOf(currentBlock.get().getSlot()).compareTo(slot) <= 0) {
         unprocessedBlocks.add(
-            ChainStorage.<BeaconBlock, PriorityQueue<BeaconBlock>>remove(this.unprocessedBlocks));
+            ChainStorage.<BeaconBlock, PriorityBlockingQueue<BeaconBlock>>remove(
+                this.unprocessedBlocks));
       } else {
         unproccesedBlocksLeft = false;
       }
