+++ validator/build.gradle
@@ -1,0 +1,1 @@
+jar { enabled = false }
