+++ validator/client/src/main/java/tech/pegasys/artemis/validator/client/CommitteeAssignmentTuple.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.validatorclient;
+package tech.pegasys.artemis.validator.client;
 
 import java.util.List;
