+++ validator/client/src/main/java/tech/pegasys/artemis/validator/client/ValidatorClient.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.artemis.validatorclient;
+package tech.pegasys.artemis.validator.client;
 
 import static tech.pegasys.artemis.datastructures.Constants.EPOCH_LENGTH;
 import static tech.pegasys.artemis.datastructures.util.BeaconStateUtil.get_crosslink_committees_at_slot;
