+++ validator/coordinator/build.gradle
@@ -1,0 +1,20 @@
+jar {
+  baseName 'validator-coordinator'
+  manifest {
+    attributes('Implementation-Title': baseName,
+    'Implementation-Version': project.version)
+  }
+}
+
+dependencies {
+  implementation project(':ethereum:statetransition')
+  implementation project(':ethereum:datastructures')
+  implementation project(':util')
+  implementation project(':pow')
+
+  implementation 'net.consensys.cava:cava-kv'
+  implementation 'net.consensys.cava:cava-bytes'
+  implementation 'com.google.guava:guava'
+  implementation 'org.apache.logging.log4j:log4j-api'
+  runtime 'org.apache.logging.log4j:log4j-core'
+}
